Hello, world!
Printing something now. Wheeeee.

This has been clarifying with respect to two things about Hare's
relationship to C. Firstly the array type mumbo jump explained in
the comment in type.lisp. Secondly I have a new understanding of C
global variables and constness.

To sum up, this is legal:

const char* hello = "Hello, world!";
int f() { hello = "Fuck you!"; return puts(hello); }

the "const" refers to "char", not the pointer itself, or in other
words, "const char*" means "pointer to const char". The string
itself can't be altered (e.g. hello[0] = 'c' is illegal) but
hello itself is just a variable. In a normal C implementation this
means that there is a sort of implicit value of type const char**,
call it mhello; source references to "hello" are really "*mhello".

This is not how Hare is going to work, because variables are not
modifiable. If a variable is initialized with "Hello", it has type
pointer to array, and you can't assign a new array.
(If a variable is initialized with an integer for example, it IS
modifiable, by writing into the int*.)